Revert "refactor: reduce empty cycling deps (#35570)"
This reverts commit 0afffe55dcbab61747c3cc57cbfa1b301b986a87.
