+++ components/config-provider/renderEmpty.tsx
@@ -3,7 +3,7 @@ import Empty from '../empty';
 import type { ConfigConsumerProps } from '.';
 import { ConfigConsumer } from '.';
 
-const defaultRenderEmpty = (componentName?: string): React.ReactNode => (
+const renderEmpty = (componentName?: string): React.ReactNode => (
   <ConfigConsumer>
     {({ getPrefixCls }: ConfigConsumerProps) => {
       const prefix = getPrefixCls('empty');
@@ -19,16 +19,13 @@ const defaultRenderEmpty = (componentName?: string): React.ReactNode => (
         case 'Transfer':
         case 'Mentions':
           return <Empty image={Empty.PRESENTED_IMAGE_SIMPLE} className={`${prefix}-small`} />;
-
-        /* istanbul ignore next */
         default:
-          // Should never hit if we take all the component into consider.
           return <Empty />;
       }
     }}
   </ConfigConsumer>
 );
 
-export type RenderEmptyHandler = typeof defaultRenderEmpty;
+export type RenderEmptyHandler = typeof renderEmpty;
 
-export default defaultRenderEmpty;
+export default renderEmpty;
